Update slice test and fix e2e test error
Move actions test to the slice test.

Fix e2e test error.

Apply design on the not found page.
