+++ src/NotFoundPage.jsx
@@ -1,7 +1,20 @@
 import React from 'react';
 
+import styled from '@emotion/styled';
+
+import Container from './styles/Container';
+
+import { colors } from './designSystem';
+
+const Title = styled.h3({
+  fontSize: '1.5em',
+  color: colors.highlight,
+});
+
 export default function NotFoundPage() {
   return (
-    <p>404 Not Found</p>
+    <Container>
+      <Title>404 Not Found</Title>
+    </Container>
   );
 }
